Fixed elasticsearch on mac

+++ src/main/java/com/smartlott/backend/persistence/domain/backend/Category.java
@@ -90,6 +90,6 @@ public class Category implements Serializable{
 
     @Override
     public int hashCode() {
-        return (int) (id ^ (id >>> 32));
+        return id ^ (id >>> 32);
     }
 }
